# flashvid-platform-gin/docs/api.yaml
openapi: 3.0.3
info:
  title: FlashVid Platform API
  description: |
    FlashVid Platform 短视频社交平台 RESTful API 文档

    ## 功能模块
    - 认证模块：用户注册、登录、Token刷新
    - 用户模块：用户信息管理、作品列表、喜欢列表、收藏列表
    - 关注模块：关注、取消关注、关注列表、粉丝列表
    - 视频模块：视频详情、发布视频、删除视频、搜索视频
    - 视频流模块：推荐流、关注流、附近流
    - 互动模块：点赞、收藏、分享
    - 评论模块：评论列表、发表评论、删除评论、点赞评论
    - 话题模块：话题列表、话题详情、话题视频列表、搜索话题
    - 上传模块：获取上传凭证、上传文件
    - 音乐模块：音乐列表、搜索音乐
    - 私信模块：会话列表、对话消息、发送私信、标记已读、删除消息、未读总数
    - 播放列表模块：获取列表、创建、更新、删除、获取列表内视频、添加/移除视频

    ## 认证方式
    使用 JWT Token 认证，在请求头中添加：
    ```
    Authorization: Bearer <access_token>
    ```

    ## 错误码
    - 0: 成功
    - 10000-19999: 用户相关错误
    - 20000-29999: 视频相关错误
    - 30000-39999: 评论相关错误
    - 40000-49999: 系统错误
    - 50000-59999: 第三方服务错误
  version: 1.0.0
  contact:
    name: FlashVid Platform API Support
    email: support@flashvid.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.flashvid.com/api/v1
    description: 生产环境
  - url: https://api-staging.flashvid.com/api/v1
    description: 测试环境
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
paths:
...
  /playlists:
    get:
      tags:
        - 播放列表
      summary: 获取我的播放列表
      description: 获取当前登录用户的所有播放列表（按创建时间倒序）
      operationId: getMyPlaylists
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          playlists:
                            type: array
                            items:
                              $ref: '#/components/schemas/PlaylistInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - BearerAuth: []

    post:
      tags:
        - 播放列表
      summary: 创建播放列表
      description: 创建一个新的自定义播放列表
      operationId: createPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  maxLength: 50
                  description: 播放列表标题
                  example: 我最喜欢的视频
                description:
                  type: string
                  maxLength: 500
                  description: 描述（可选）
                  example: 收藏的精彩视频
                coverUrl:
                  type: string
                  format: uri
                  description: 封面URL（可选）
                  example: https://cdn.flashvid.com/cover/playlist.jpg
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          playlist:
                            $ref: '#/components/schemas/PlaylistInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
      security:
        - BearerAuth: []
  /playlists/{id}:
    put:
      tags:
        - 播放列表
      summary: 更新播放列表
      description: 更新播放列表的标题、描述或封面
      operationId: updatePlaylist
      parameters:
        - name: id
          in: path
          required: true
          description: 播放列表ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 50
                  example: 新标题
                description:
                  type: string
                  maxLength: 500
                  example: 新描述
                coverUrl:
                  type: string
                  format: uri
                  example: https://cdn.flashvid.com/cover/new.jpg
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []

    delete:
      tags:
        - 播放列表
      summary: 删除播放列表
      description: 删除指定播放列表及其所有视频关联
      operationId: deletePlaylist
      parameters:
        - name: id
          in: path
          required: true
          description: 播放列表ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
  /playlists/{id}/videos:
    get:
      tags:
        - 播放列表
      summary: 获取播放列表内的视频
      description: 游标分页获取指定播放列表内的视频列表（按加入时间倒序）
      operationId: getPlaylistVideos
      parameters:
        - name: id
          in: path
          required: true
          description: 播放列表ID
          schema:
            type: integer
            format: int64
        - name: cursor
          in: query
          required: false
          description: 游标（上一页最后一条记录的 created_at，格式 "yyyy-MM-dd HH:mm:ss"，首次请求不传）
          schema:
            type: string
            example: "2026-08-01 12:00:00"
        - name: count
          in: query
          required: false
          description: 每页数量，默认 20，最大 50
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          playlist:
                            $ref: '#/components/schemas/PlaylistInfo'
                          videos:
                            type: array
                            items:
                              $ref: '#/components/schemas/Video'
                          nextCursor:
                            type: string
                            nullable: true
                            description: 下一页游标，为 null 表示没有更多数据
                            example: "2026-07-20T08:30:00Z"
                          hasMore:
                            type: boolean
                            description: 是否还有更多数据
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []

    post:
      tags:
        - 播放列表
      summary: 添加视频到播放列表
      description: 将指定视频添加到播放列表（幂等，重复添加返回已在列表错误）
      operationId: addVideoToPlaylist
      parameters:
        - name: id
          in: path
          required: true
          description: 播放列表ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - videoId
              properties:
                videoId:
                  type: integer
                  format: int64
                  description: 视频ID
                  example: 12345
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 视频已在播放列表中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
      security:
        - BearerAuth: []
  /playlists/{id}/videos/{videoId}:
    delete:
      tags:
        - 播放列表
      summary: 从播放列表移除视频
      description: 将指定视频从播放列表中移除
      operationId: removeVideoFromPlaylist
      parameters:
        - name: id
          in: path
          required: true
          description: 播放列表ID
          schema:
            type: integer
            format: int64
        - name: videoId
          in: path
          required: true
          description: 视频ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 移除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证，格式：Bearer <token>
...
  schemas:
    Response:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 响应码，0表示成功
          example: 0
        message:
          type: string
          description: 响应消息
          example: success
        data:
          type: object
          nullable: true
          description: 响应数据
...
    Video:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 视频ID
          example: 12345
        title:
          type: string
          description: 视频标题
          example: 我的第一个短视频
        description:
          type: string
          description: 视频描述
          example: 这是一个测试视频
        coverUrl:
          type: string
          format: uri
          description: 封面URL
          example: https://cdn.flashvid.com/cover/12345.jpg
        videoUrl:
          type: string
          format: uri
          description: 视频URL
          example: https://cdn.flashvid.com/video/12345.mp4
        duration:
          type: integer
          description: 视频时长（秒）
          example: 30
        width:
          type: integer
          description: 视频宽度
          example: 1080
        height:
          type: integer
          description: 视频高度
          example: 1920
        musicId:
          type: integer
          format: int64
          description: 背景音乐ID
          example: 456
        city:
          type: string
          description: 城市
          example: 深圳
        topics:
          type: array
          items:
            type: string
          description: 话题标签
          example: ['旅行', '美食']
        author:
          type: object
          description: 作者信息
          properties:
            id:
              type: integer
              format: int64
              example: 123
            username:
              type: string
              example: author
            nickname:
              type: string
              example: 作者
            avatar:
              type: string
              format: uri
              example: https://cdn.flashvid.com/avatar/123.jpg
        stats:
          type: object
          description: 统计信息
          properties:
            viewCount:
              type: integer
              description: 播放数
              example: 10000
            likeCount:
              type: integer
              description: 点赞数
              example: 1000
            commentCount:
              type: integer
              description: 评论数
              example: 100
            shareCount:
              type: integer
              description: 分享数
              example: 50
            favoriteCount:
              type: integer
              description: 收藏数
              example: 200
        publishedAt:
          type: string
          description: 发布时间（格式：2006-01-02 15:04:05）
          example: '2024-01-01 12:00:00'
...
    PlaylistInfo:
      type: object
      description: 播放列表信息
      properties:
        id:
          type: integer
          format: int64
          description: 播放列表ID
          example: 1
        title:
          type: string
          description: 标题
          example: 我最喜欢的视频
        description:
          type: string
          description: 描述
          example: 收藏的精彩视频
        coverUrl:
          type: string
          format: uri
          description: 封面URL
          example: https://cdn.flashvid.com/cover/playlist.jpg
        videoCount:
          type: integer
          format: int32
          description: 视频数量
          example: 12
        createdAt:
          type: string
          description: 创建时间（格式：2006-01-02 15:04:05）
          example: '2024-01-01 12:00:00'
...
  responses:
...
    Unauthorized:
      description: 未登录或Token失效
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Response'
              - type: object
                properties:
                  code:
                    example: 40001
                  message:
                    example: 未登录或Token失效
                  data:
                    example: null
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Response'
              - type: object
                properties:
                  code:
                    example: 40003
                  message:
                    example: 无权限访问
                  data:
                    example: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Response'
              - type: object
                properties:
                  code:
                    example: 40004
                  message:
                    example: 资源不存在
                  data:
                    example: null
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Response'
              - type: object
                properties:
                  code:
                    example: 42200
                  message:
                    example: 请求参数验证失败
                  data:
                    type: object
                    properties:
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                              example: username
                            message:
                              type: string
                              example: 用户名不能为空
...